requirements.txt
fastapi==0.112.2 # temporary fix of HuggingFace https://discuss.huggingface.co/t/new-gradio-space-connection-errored-out/105509/3
transformers==4.36.2
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
numpy==1.24.4
scipy==1.10.1
scikit-image==0.21.0
opencv-python==4.7.0.72
pillow==9.4.0
diffusers==0.25.0
transformers==4.36.2
accelerate==0.26.1
matplotlib==3.7.4
tqdm==4.64.1
config==0.5.1
einops==0.7.0
onnxruntime==1.16.2
basicsr
av
fvcore
cloudpickle
omegaconf
pycocotools
huggingface_hub==0.25.2
pydantic==2.8.2

# for libcom library - build is failing so commenting it again
# libcom

# For Local
gradio==4.24.0
spaces==0.28.3